backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi-cors==0.0.6
pillow==10.1.0
aiofiles==23.2.1